Switch Ruby implementation to Matrix
Rubyists rarely use recursion and TCO is even disabled by default, so this is awkward for Ruby. This is an alternative Ruby implementation using Matrix rather than recursion. It's hundreds of times faster but may not be in the spirit of the competition. I thought I'd propose it anyways, since recursion is not how Rubyists would typically tackle this one.

Maybe this could be an alternative, optimized version, if it's not allowed by the rules? I just wanted to put it out there. Thanks!