Make loops comparable in go and c
The clang compiler performs an optimization called Scalar Replacement of
Aggregates on the C code which essentially means that it rewrites the c code to
use a local variable (which is stored in a register) just for iterating through
the inner loop with 100,000 sums of `j%u`, and then assigns it back to the array
of int32 values.

The go complier doesn't currently perform the same optimization. But, it's
trivial to rewrite the code manually to perform the same optimization. So, it's
not exactly the same code, but using only the features of the go programming
language, this code now performs as well as the equivalent c code on my mac.