Optimize julia loop implementation
Encapsulating the code in a function allows it to be compiled. The basic compilation unit in
Julia is a function, so any code at the top level is not compiled.

"@inbounds" macros were removed as they didn't make a difference for performance in this
case.

All other changes were stylistic to make the code more idiomatic.