convert run.sh to run
Importantly, it puts all the possible commands into a single hyperfine
invocation, which can help you get a relative comparison in the final
report (see below). Also, use a normal number of runs and warmups to get
more statistically significant results.

The "runnable" test also improves: it checks that the program exists &&
that the runner program (which may be the same) exists and is
executable. That means if you don't have `miniruby` installed, for
example, it should be skipped.

The use of Zsh is for portability with associative arrays (macOS's
/bin/bash is still too old on most systems).

Other noteworthy changes:
- dropped Hare because it didn't appear to work with existing script
- dropped commented out programs
- "descriptions" lost, but could be re-added in a separate parallel
  array
- kept Clojure program filename, though it doesn't look right to me
- common-lisp file check includes .lisp suffix (in run.sh, only the
  command had the suffix)
- Clojure runner construction with -Spath might not work…

Example run with just Racket and C:

    # ../run
    Benchmark 1: racket/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
      Time (mean ± σ):      5.327 s ±  0.197 s    [User: 4.991 s, System: 0.112 s]
      Range (min … max):    5.118 s …  5.700 s    10 runs

    Benchmark 2: c/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
      Time (mean ± σ):     836.5 ms ±   3.8 ms    [User: 832.3 ms, System: 1.9 ms]
      Range (min … max):   832.5 ms … 843.3 ms    10 runs

    Summary
      c/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa... ran
        6.37 ± 0.24 times faster than racket/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...

Makes-progress-on: https://github.com/bddicken/languages/issues/320